feat: Create component Products
